+++ src/main/scala/com/martomate/hexacraft/renderer/VertexData.scala
@@ -1,0 +1,9 @@
+package com.martomate.hexacraft.renderer
+
+import java.nio.ByteBuffer
+
+trait VertexData {
+  def bytesPerVertex: Int
+
+  def fill(buf: ByteBuffer): Unit
+}
