+++ src/main/scala/com/martomate/hexacraft/world/coord/integer/BlockRelChunk.scala
@@ -1,7 +1,7 @@
 package com.martomate.hexacraft.world.coord.integer
 
 import com.martomate.hexacraft.util.CylinderSize
-import com.martomate.hexacraft.world.block.state.BlockState
+import com.martomate.hexacraft.world.coord.NeighborOffsets
 
 object BlockRelChunk {
   def apply(x: Int, y: Int, z: Int)(implicit cylSize: CylinderSize): BlockRelChunk = BlockRelChunk((x & 0xf) << 8 | (y & 0xf) << 4 | (z & 0xf))
@@ -15,7 +15,7 @@ case class BlockRelChunk(private val _value: Int)(implicit val cylSize: Cylinder
   def offset(x: Int, y: Int, z: Int): BlockRelChunk = BlockRelChunk(cx + x, cy + y, cz + z)
 
   def onChunkEdge(side: Int): Boolean = {
-    val off = BlockState.neighborOffsets(side)
+    val off = NeighborOffsets(side)
     val xx = cx + off._1
     val yy = cy + off._2
     val zz = cz + off._3
@@ -23,7 +23,7 @@ case class BlockRelChunk(private val _value: Int)(implicit val cylSize: Cylinder
   }
 
   def neighbor(side: Int): BlockRelChunk = {
-    val off = BlockState.neighborOffsets(side)
+    val off = NeighborOffsets(side)
     val xx = cx + off._1
     val yy = cy + off._2
     val zz = cz + off._3
